Add file uris to cache api
